nhl_projection_phase1_reset_audit/requirements.txt
pandas>=2.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.1
pyarrow>=18.0.0
google-api-python-client>=2.169.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
